1. Moves core/src/test => core/test 2.Update TODOs
