+++ core/test/config_test.ts
@@ -1,8 +1,7 @@
-import { asserts } from '../../deps_dev.ts';
-import * as config from '../config.ts';
-import * as logging from '../logging.ts';
-import { ToInstance } from '../record.ts';
-import { load } from '../yaml.ts';
+import { asserts } from '../deps_dev.ts';
+import { config, logging, record, yaml } from '../mod.ts';
+const { ToInstance } = record;
+const { load } = yaml;
 
 // current directory for this file
 const dirname = new URL('.', import.meta.url).pathname;
@@ -12,7 +11,7 @@ const log = logging.Category('config.test');
 const testAutomateConfig = function(plain: any) {
   const cfg = ToInstance(config.AutomateConfig, plain);
   cfg.convertTypes();
-  console.log(cfg);
+  // console.log(cfg);
   const deps = cfg.dependencies;
   if (deps !== undefined) {
     console.log('Log dependencies');
