+++ core/test/fixture/Automate.yaml
