+++ core/test/fixture/yaml-merge-1.yaml
