+++ core/test/fixture/yaml-merge-2.yaml
