+++ core/test/fixture/yaml-merge-3.yaml
