+++ core/test/fixture/yaml-test-multiple.yaml
