+++ core/test/fixture/yaml-test.yaml
