+++ core/test/package_test.ts
@@ -1,0 +1,16 @@
+import { asserts } from '../deps_dev.ts';
+import { pkg } from '../mod.ts';
+
+const dirname = new URL('.', import.meta.url).pathname;
+
+const pkgRecipeName1File =
+  `${dirname}/workspace-example/recipe/name1/Automate.yaml`;
+
+Deno.test(
+  async function testPackageFromPath() {
+    const pkgRecipeName1 = await pkg.Package.fromPath(pkgRecipeName1File);
+
+    // console.log(pkgRecipeName1.toObject());
+    // console.log(pkgRecipeName1);
+  },
+);
