+++ core/test/workspace-example/Automate.yaml
