+++ core/test/workspace-example/deno.jsonc
