+++ core/test/workspace-example/provider/name1/Automate.yaml
