+++ core/test/workspace-example/provider/name1/mod_test.ts
