+++ core/test/workspace-example/provider/name2/Automate.yaml
