+++ core/test/workspace-example/provider/name2/mod.ts
@@ -1,11 +1,11 @@
-import { logging, provider } from '../../../../../mod.ts';
+import { logging, provider } from '../../../../mod.ts';
 
 // create logger
-const log = logging.Category('name3');
+const log = logging.Category('name2');
 
 type Values = Record<string, unknown>;
 
-class ProviderName3 implements provider.Provider {
+class ProviderName2 implements provider.Provider {
   constructor() {}
 
   // deno-lint-ignore require-await
@@ -17,6 +17,6 @@ class ProviderName3 implements provider.Provider {
 
 // deno-lint-ignore require-await
 export const initializeProvider = async (): Promise<provider.Provider> => {
-  const instance = new ProviderName3();
+  const instance = new ProviderName2();
   return Promise.resolve(instance);
 };
