+++ core/test/workspace-example/provider/name2/mod_test.ts
