+++ core/test/workspace-example/provider/name3/Automate.yaml
