+++ core/test/workspace-example/provider/name3/mod_test.ts
