+++ core/test/workspace-example/recipe/name1/Automate.yaml
