+++ core/test/workspace-example/recipe/name2/Automate.yaml
