+++ core/test/yaml_test.ts
@@ -1,6 +1,8 @@
-import { asserts } from '../../deps_dev.ts';
-import { ToInstance, Type } from '../record.ts';
-import { load, loadAll, mergeLoad } from '../yaml.ts';
+import { asserts } from '../deps_dev.ts';
+import { record, yaml } from '../mod.ts';
+
+const { ToInstance, Type } = record;
+const { load, loadAll, mergeLoad } = yaml;
 
 // current directory for this file
 const dirname = new URL('.', import.meta.url).pathname;
